Implemented Song Filter and Switched to ListView
- Many UI changes and bug fixes
- Help form added

+++ MusicSorter/MainForm.cs
@@ -10,25 +10,45 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using System.Windows.Forms;
+using MusicSorter.Helpers;
+using MusicSorter.Forms;
 
 namespace MusicSorter
 {
     public partial class MainForm : Form
     {
         private bool Playing { get; set; }
-        private bool ValidSong { get; set; } = false;
+        private string CurrentSong { get; set; }
         private WaveOutEvent OutputDevice { get; set; }
         private AudioFileReader AudioFile { get; set; }
+        private SongRater songRater { get; set; }
+
+
+        private int AcceptableRating = 5;
+        private int GreatRating = 8;
 
         public MainForm()
         {
             InitializeComponent();
         }
 
+        private void MainForm_Load(object sender, EventArgs e)
+        {
+            try
+            {
+                songRater = new SongRater();
+                textBoxMusicFolder.Text = songRater.PreviousFolder;
+                PopulateSongs(textBoxMusicFolder.Text);
+            }
+            catch (Exception ex)
+            {
+                MessageBox.Show(ex.Message);
+            }
+        }
+
         //TODO track bar
-        //TODO help shortcuts
         //TODO icon
-        //TODO stop deleted song from playing twice (keep track of all words and then if 100% match then stop it), show as yellow in right and is skipped
+        //TODO reponse label for successfully saving and deleting
 
         private void buttonBrowse_Click(object sender, EventArgs e)
         {
@@ -53,6 +73,7 @@ namespace MusicSorter
             try
             {
                 PopulateSongs(textBoxMusicFolder.Text);
+                songRater.PreviousFolder = textBoxMusicFolder.Text;
             }
             catch (Exception ex)
             {
@@ -64,7 +85,7 @@ namespace MusicSorter
         {
             try
             {
-                TogglePlay();
+                TogglePlay(!Playing);
             }
             catch (Exception ex)
             {
@@ -72,37 +93,6 @@ namespace MusicSorter
             }
         }
 
-        private void listBoxSongs_SelectedIndexChanged(object sender, EventArgs e)
-        {
-            try
-            {
-                LoadSong();
-            }
-            catch (Exception ex)
-            {
-                MessageBox.Show(ex.Message);
-            }
-        }
-
-        private void MainForm_KeyPress(object sender, KeyPressEventArgs e)
-        {
-            try
-            {
-                int i;
-                if (int.TryParse(e.KeyChar.ToString(), out i))
-                {
-                    MoveSongPosition(i);
-                }
-                else if (e.KeyChar == ' ')
-                {
-                    TogglePlay();
-                }
-            }
-            catch (Exception ex)
-            {
-                MessageBox.Show(ex.Message);
-            }
-        }
 
         private void buttonNext_Click(object sender, EventArgs e)
         {
@@ -131,23 +121,48 @@ namespace MusicSorter
         {
             try
             {
-                switch (e.KeyCode)
+                if ((e.KeyCode >= Keys.D0 && e.KeyCode <= Keys.D9) ||
+                    (e.KeyCode >= Keys.NumPad0 && e.KeyCode <= Keys.NumPad9))
                 {
-                    case Keys.Down:
-                        NextSong();
-                        break;
-                    case Keys.Up:
-                        PrevSong();
-                        break;
-                    case Keys.Left:
-                        SkipSongPosition(-1);
-                        break;
-                    case Keys.Right:
-                        SkipSongPosition(1);
-                        break;
+                    int i;
+                    if (int.TryParse(e.KeyCode.ToString().Substring(1), out i))
+                    {
+                        if (e.Modifiers == Keys.Control)
+                        {
+                            if (listViewSongs.SelectedIndices != null && listViewSongs.SelectedIndices.Count > 0)
+                            {
+                                UpdateRating(listViewSongs.SelectedIndices[0], i);
+                            }                            
+                        }
+                        else
+                        {
+                            MoveSongPosition(i);
+                        }
+                    }
+                }
+                else
+                {
+                    switch (e.KeyCode)
+                    {
+                        case Keys.Down:
+                            NextSong();
+                            break;
+                        case Keys.Up:
+                            PrevSong();
+                            break;
+                        case Keys.Left:
+                            SkipSongPosition(-1);
+                            break;
+                        case Keys.Right:
+                            SkipSongPosition(1);
+                            break;
+                        case Keys.Space:
+                            TogglePlay(!Playing);
+                            break;
 
-                    default:
-                        break;
+                        default:
+                            break;
+                    }
                 }
             }
             catch (Exception ex)
@@ -156,16 +171,37 @@ namespace MusicSorter
             }
         }
 
-        private void buttonDelete_Click(object sender, EventArgs e)
+        private void buttonSave_Click(object sender, EventArgs e)
         {
             try
             {
-                var selectedSong = Convert.ToString(listBoxSongs.SelectedItem);
-                var fileLocation = Path.Combine(textBoxMusicFolder.Text, selectedSong);
-                if (File.Exists(fileLocation))
+                songRater.SaveSongRatings();
+                songRater.SongsRatingsUpdated = false;
+            }
+            catch (Exception ex)
+            {
+                MessageBox.Show(ex.Message);
+            }
+        }
+
+        private void buttonFilter_Click(object sender, EventArgs e)
+        {
+            try
+            {
+                for (int i = 0; i < listViewSongs.Items.Count; i++)
                 {
-                    NextSong();
-                    FileSystem.DeleteFile(fileLocation, UIOption.OnlyErrorDialogs, RecycleOption.SendToRecycleBin);
+                    var formattedName = GetFormattedName(listViewSongs.Items[i].Text);
+                    if (songRater.SongRatings.ContainsKey(formattedName)
+                        && songRater.SongRatings[formattedName] < AcceptableRating)
+                    {
+                        var selectedSong = listViewSongs.Items[i].Text;
+                        var fileLocation = Path.Combine(textBoxMusicFolder.Text, selectedSong);
+                        if (File.Exists(fileLocation))
+                        {
+                            NextSong();
+                            FileSystem.DeleteFile(fileLocation, UIOption.OnlyErrorDialogs, RecycleOption.SendToRecycleBin);
+                        }
+                    }
                 }
             }
             catch (Exception ex)
@@ -174,14 +210,77 @@ namespace MusicSorter
             }
         }
 
-        private void listBoxSongs_KeyDown(object sender, KeyEventArgs e)
+        private void listViewSongs_SelectedIndexChanged(object sender, EventArgs e)
         {
-            if (e.KeyCode == Keys.Right
-                || e.KeyCode == Keys.Left
-                || e.KeyCode == Keys.Up
-                || e.KeyCode == Keys.Down)
+            try
             {
-                e.Handled = true;
+                if (listViewSongs.SelectedItems != null && listViewSongs.SelectedItems.Count > 0)
+                {
+                    LoadSong(listViewSongs.SelectedItems[0].Text);
+                    TogglePlay(true);
+                }
+            }
+            catch (Exception ex)
+            {
+                MessageBox.Show(ex.Message);
+            }
+        }
+
+        private void MainForm_FormClosing(object sender, FormClosingEventArgs e)
+        {
+            try
+            {
+                if (songRater.SongsRatingsUpdated)
+                {
+                    var dr = MessageBox.Show("Do you want to save your unsaved rating changes?", "Confirmation", MessageBoxButtons.YesNoCancel);
+                    if (dr == DialogResult.Yes)
+                    {
+                        songRater.SaveSongRatings();
+                    }
+                    else if (dr == DialogResult.Cancel)
+                    {
+                        e.Cancel = true;
+                    }
+                }
+            }
+            catch (Exception ex)
+            {
+                MessageBox.Show(ex.Message);
+            }
+        }
+
+        private void listViewSongs_KeyDown(object sender, KeyEventArgs e)
+        {
+            e.SuppressKeyPress = true;
+        }
+
+        private void buttonHelp_Click(object sender, EventArgs e)
+        {
+            HelpForm form = new HelpForm();
+            form.ShowDialog();
+        }
+
+        private void UpdateRating(int pos, int rating)
+        {
+            var formattedName = GetFormattedName(CurrentSong);
+            songRater.UpdateRating(formattedName, rating);
+            songRater.SongsRatingsUpdated = true;
+            UpdateColor(pos, rating);
+        }
+
+        private void UpdateColor(int pos, int rating)
+        {
+            if (rating >= GreatRating)
+            {
+                listViewSongs.Items[pos].BackColor = Color.Gold;
+            }
+            else if (rating >= AcceptableRating)
+            {
+                listViewSongs.Items[pos].BackColor = Color.LightGreen;
+            }
+            else
+            {
+                listViewSongs.Items[pos].BackColor = Color.Salmon;
             }
         }
 
@@ -192,62 +291,137 @@ namespace MusicSorter
                 //string[] files = System.IO.Directory.GetFiles(folder, "*.mp3");
                 string[] files = System.IO.Directory.GetFiles(folder);
 
-                listBoxSongs.Items.Clear();
-                listBoxSongs.Items.AddRange(files);
-
-                if (listBoxSongs.Items.Count > 0)
+                listViewSongs.Items.Clear();
+                foreach (var file in files)
                 {
-                    listBoxSongs.SelectedIndex = 0;
-                    LoadSong();
+                    listViewSongs.Items.Add(Path.GetFileName(file));
+                }
+
+                for (int i = 0; i < listViewSongs.Items.Count; i++)
+                {
+                    var songFileName = listViewSongs.Items[i].Text;
+                    var formattedName = GetFormattedName(songFileName);
+
+                    if (songRater.SongRatings.ContainsKey(formattedName))
+                    {
+                        var rating = songRater.SongRatings[formattedName];
+                        UpdateColor(i, rating);
+                    }
+                }
+
+                //find first acceptable song
+                if (listViewSongs.Items.Count > 0)
+                {
+                    for (int i = 0; i < listViewSongs.Items.Count; i++)
+                    {
+                        var formattedName = GetFormattedName(listViewSongs.Items[i].Text);
+                        if (!songRater.SongRatings.ContainsKey(formattedName)
+                            || songRater.SongRatings[formattedName] >= AcceptableRating)
+                        {
+                            listViewSongs.Items[i].Selected = true;
+                            LoadSong(listViewSongs.SelectedItems[0].Text);
+                            TogglePlay(true);
+                            break;
+                        }
+                    }
                 }
             }
         }
 
-        private void LoadSong()
+        private string GetFormattedName(string songFileName)
         {
-            var selectedSong = Convert.ToString(listBoxSongs.SelectedItem);
+            var songFileLocation = GetSongFileLocation(songFileName);
+            if (!string.IsNullOrWhiteSpace(songFileLocation))
+            {
+                using (var file = TagLib.File.Create(songFileLocation))
+                {
+                    if (file.Tag.Performers.Length >= 1
+                        && !string.IsNullOrWhiteSpace(file.Tag.Title))
+                    {
+                        var performers = file.Tag.Performers.Where(x => !string.IsNullOrEmpty(x)).ToArray();
+                        Array.Sort(performers);
+                        for (int i = 0; i < performers.Length; i++)
+                        {
+                            performers[i] = performers[i].Trim();
+                        }
+                        return songRater.FormatSongName(string.Join(", ", performers), file.Tag.Title);
+                    }
+                }
+            }
+            return songRater.FormatSongName(songFileName);
+        }
+
+        private void LoadSong(string selectedSong)
+        {
+            var fileLocation = GetSongFileLocation(selectedSong);
+            if (string.IsNullOrWhiteSpace(fileLocation)) return;
+
+            textBoxCurrentSong.Text = selectedSong;
+
+            try
+            {
+                if (OutputDevice != null)
+                {
+                    OutputDevice.Dispose();
+                }
+                if (AudioFile != null)
+                {
+                    AudioFile.Dispose();
+                }
+
+                AudioFile = new AudioFileReader(fileLocation);
+                OutputDevice = new WaveOutEvent();
+                OutputDevice.Init(AudioFile);
+                OutputDevice.PlaybackStopped += OutputDevice_PlaybackStopped;
+                CurrentSong = selectedSong;
+
+                var file = TagLib.File.Create(fileLocation);
+                if (file.Tag.Pictures.Length >= 1)
+                {
+                    var bin = (byte[])(file.Tag.Pictures[0].Data.Data);
+                    var length = Math.Min(pictureBoxAlbumArt.Width, pictureBoxAlbumArt.Height);
+                    pictureBoxAlbumArt.Image = Image.FromStream(new MemoryStream(bin)).GetThumbnailImage(length, length, null, IntPtr.Zero);
+                }
+                else
+                {
+                    pictureBoxAlbumArt.Image = null;
+                }
+            }
+            catch
+            {
+                textBoxCurrentSong.Text = "Audio file type is not supported.";
+                CurrentSong = string.Empty;
+            }
+        }
+
+        private void OutputDevice_PlaybackStopped(object sender, StoppedEventArgs e)
+        {
+            if (AudioFile.Position == AudioFile.Length)
+            {
+                NextSong();
+            }
+        }
+
+        private string GetSongFileLocation(string selectedSong)
+        {
             var fileLocation = Path.Combine(textBoxMusicFolder.Text, selectedSong);
             if (File.Exists(fileLocation))
             {
-                textBoxCurrentSong.Text = selectedSong;
-
-                try
-                {
-                    if (OutputDevice != null)
-                    {
-                        OutputDevice.Dispose();
-                    }
-                    if (AudioFile != null)
-                    {
-                        AudioFile.Dispose();
-                    }
-
-                    AudioFile = new AudioFileReader(fileLocation);
-                    OutputDevice = new WaveOutEvent();
-                    OutputDevice.Init(AudioFile);
-                    ValidSong = true;
-
-                    var file = TagLib.File.Create(fileLocation);
-                    if (file.Tag.Pictures.Length >= 1)
-                    {
-                        var bin = (byte[])(file.Tag.Pictures[0].Data.Data);
-                        var length = Math.Min(pictureBoxAlbumArt.Width, pictureBoxAlbumArt.Height);
-                        pictureBoxAlbumArt.Image = Image.FromStream(new MemoryStream(bin)).GetThumbnailImage(length, length, null, IntPtr.Zero);
-                    }
-                }
-                catch
-                {
-                    textBoxCurrentSong.Text = "Audio file type is not supported.";
-                    ValidSong = false;
-                }
+                return fileLocation;
+            }
+            else
+            {
+                return string.Empty;
             }
         }
 
-        private void TogglePlay()
+        private void TogglePlay(bool play)
         {
-            Playing = !Playing;
+            Playing = play;
 
-            if (ValidSong)
+            if (!string.IsNullOrWhiteSpace(CurrentSong)
+                && OutputDevice != null
+                && AudioFile != null)
             {
                 if (Playing)
                 {
@@ -266,7 +440,9 @@ namespace MusicSorter
         {
             var segment = AudioFile.Length / 10;
             AudioFile.Position = segment * i;
+            TogglePlay(true);
         }
+
         private void SkipSongPosition(int i)
         {
             var segment = AudioFile.Length / 100;
@@ -279,19 +455,37 @@ namespace MusicSorter
 
         private void NextSong()
         {
-            if (listBoxSongs.SelectedIndex < listBoxSongs.Items.Count - 1)
+            if (listViewSongs.SelectedIndices != null)
             {
-                listBoxSongs.SelectedIndex++;
-                OutputDevice.Play();
+                int i = listViewSongs.SelectedIndices[0];
+                while (++i < listViewSongs.Items.Count)
+                {
+                    var formattedName = GetFormattedName(listViewSongs.Items[i].Text);
+                    if (!songRater.SongRatings.ContainsKey(formattedName)
+                        || songRater.SongRatings[formattedName] >= AcceptableRating)
+                    {
+                        listViewSongs.Items[i].Selected = true;
+                        break;
+                    }
+                }
             }
         }
 
         private void PrevSong()
         {
-            if (listBoxSongs.SelectedIndex > 0)
+            if (listViewSongs.SelectedIndices != null)
             {
-                listBoxSongs.SelectedIndex--;
-                OutputDevice.Play();
+                int i = listViewSongs.SelectedIndices[0];
+                while (--i >= 0)
+                {
+                    var formattedName = GetFormattedName(listViewSongs.Items[i].Text);
+                    if (!songRater.SongRatings.ContainsKey(formattedName)
+                        || songRater.SongRatings[formattedName] >= AcceptableRating)
+                    {
+                        listViewSongs.Items[i].Selected = true;
+                        break;
+                    }
+                }
             }
         }
     }
